refactor: extract git operations to gitutil package
Create new gitutil package to centralize git CLI operations that work
around go-git v5 bugs. This provides a single location for git operations
that need to use native git commands for correctness.

New package includes:
- OpenRepository: opens repo with worktree support
- GetWorktreePath, IsInsideWorktree, GetMainRepoRoot, GetGitCommonDir
- CheckoutBranch, HardResetWithProtection (go-git bug workarounds)
- ValidateBranchName, HasUncommittedChanges, GetConfigValue
- FetchBranch, CreateLocalBranchFromRemote

The existing functions in cli/git_operations.go and strategy/common.go
are marked as deprecated and now delegate to gitutil. This allows
gradual migration of callers to the new package.

Entire-Checkpoint: ba9a1bd49e00
